requirements.txt
toolbox @ git+https://github.com/Valcrist/toolbox.git
numpy
pillow
imagehash
scikit-image
